Lee Elliott:
here's the A-10 update.

I don't know if you've had a chance to look at what Alexis has
done in the A-10 cockpit but it looks pretty good to me  :)

We're still working out the correct eye/viewpoint position - we
figure that three things must come together for it to be
correct: the approach AoA, the angle of depression of the cannon
and the viewpoint.  With the correct approach AoA we must be
able to see the runway and with the correct depression of the
cannon, the shells should be aiming at the right point on the
hud, as seen from the viewpoint.  I think it's coming together.

There seems to be a bug in the view handling code though - if you
pan the view using the mouse and then reset it via a left-click,
the -12 degree pitch-down we use in the cockpit view also seems
to get applied to the other (lookat) views e.g. chase and
helicopter.

I think it's something to with the way that settings are swapped
between the individual view configs and the current view.
Something else to be looked at...