- Major update:
DavidB21: Fuel System and Engine/APU related stuff.

- Flaps lever animation near throttle.
- Real A-10 fuel pumps logic implemented (both main and wing).
- Fuel pumps will not provide enough pressure during negative G flight.
Engine have sufficient fuel in his collector for 10 seconds operation at
maximum power.
- In case of main pump failure, the affected engine will suction-feed
from the tank for all power settings up to an altitude of 10,000ft on a
standard day (29.92 InHG pressure at sea-level).
- In case of wing pump failure, fuel will gravity feed to its associated
main tank if the main tank fuel level is below 600 lbs.
- Real fuel valves logic implemented (tank gate, cross feed, external
tanks and internal fill disable valves).
- Air-Air refuel slipway door is now powered by the right hydraulic system.
- Nasal function (see A10fuel.aar_reset_button()) to reset the air-air
refuel system from "disconnected" state to "ready" state without having
to close the recover lever. This function could be tied to a joystick
button.
- New APU code. APU starts can be made up to an altitude of 15,000ft on
a standard day. APU pressure output will be sufficient to start an
engine up to an engine of 10,000ft on a standard day.
- APU generator now cooling the APU. During ground operation, if the APU
EGT exceed 720°C it will be automatically shutdown. If temperature
exceed 850°C the APU is killed.
- Right and left hydraulic pressures tied to the engines core speed.
- Throttle have now an OFF and IDLE positions. You have to click on a
hotspot in order to move the throttle from OFF to IDLE and reciprocally.
- Real engines autostart logic and engine start cycle implemented (like
real A-10 block 45 and superior).
- "NORM" and "MOTOR" engine operate switchs logic implemented. Switches
are spring-loaded from "IGN" to "NORM" position.
- "NORM" engine operate switch is used for normal operation and
autostart. "MOTOR" is used for air-purging of excessive fuel in the
collector tanks (if throttle is in OFF position).
- New engine start system: now engine start require low pressure from
the APU or cross bleed air from the other engine (85% core rpm minimum).


Alexis Bory: Pitch SAS (Stability Augmentation System)

This the first part of the A-10 SAS. (more to come)

- Panel with Pitch SAS switches (right of the throttles). Try a 30° dive
with the Pitch SAS off.
- Pitch SAS smooths the elevator stick input (function of time).
- Pitch SAS compress and damps the elevator output when positive (diving) and
AoA < 2°.
- Removed the movable ballast wich was a dirty workaround.